+++ twilio/rest/conversations/v1/address_configuration.py
@@ -32,13 +32,14 @@ class AddressConfigurationList(ListResource):
         self._solution = {}
         self._uri = '/Configuration/Addresses'.format(**self._solution)
 
-    def stream(self, limit=None, page_size=None):
+    def stream(self, type=values.unset, limit=None, page_size=None):
         """
         Streams AddressConfigurationInstance records from the API as a generator stream.
         This operation lazily loads records as efficiently as possible until the limit
         is reached.
         The results are returned as a generator, so this operation is memory efficient.
 
+        :param unicode type: The type of address configuration.
         :param int limit: Upper limit for the number of records to return. stream()
                           guarantees to never return more than limit.  Default is no limit
         :param int page_size: Number of records to fetch per request, when not set will use
@@ -51,16 +52,17 @@ class AddressConfigurationList(ListResource):
         """
         limits = self._version.read_limits(limit, page_size)
 
-        page = self.page(page_size=limits['page_size'], )
+        page = self.page(type=type, page_size=limits['page_size'], )
 
         return self._version.stream(page, limits['limit'])
 
-    def list(self, limit=None, page_size=None):
+    def list(self, type=values.unset, limit=None, page_size=None):
         """
         Lists AddressConfigurationInstance records from the API as a list.
         Unlike stream(), this operation is eager and will load `limit` records into
         memory before returning.
 
+        :param unicode type: The type of address configuration.
         :param int limit: Upper limit for the number of records to return. list() guarantees
                           never to return more than limit.  Default is no limit
         :param int page_size: Number of records to fetch per request, when not set will use
@@ -71,14 +73,15 @@ class AddressConfigurationList(ListResource):
         :returns: Generator that will yield up to limit results
         :rtype: list[twilio.rest.conversations.v1.address_configuration.AddressConfigurationInstance]
         """
-        return list(self.stream(limit=limit, page_size=page_size, ))
+        return list(self.stream(type=type, limit=limit, page_size=page_size, ))
 
-    def page(self, page_token=values.unset, page_number=values.unset,
-             page_size=values.unset):
+    def page(self, type=values.unset, page_token=values.unset,
+             page_number=values.unset, page_size=values.unset):
         """
         Retrieve a single page of AddressConfigurationInstance records from the API.
         Request is executed immediately
 
+        :param unicode type: The type of address configuration.
         :param str page_token: PageToken provided by the API
         :param int page_number: Page Number, this value is simply for client state
         :param int page_size: Number of records to return, defaults to 50
@@ -86,7 +89,12 @@ class AddressConfigurationList(ListResource):
         :returns: Page of AddressConfigurationInstance
         :rtype: twilio.rest.conversations.v1.address_configuration.AddressConfigurationPage
         """
-        data = values.of({'PageToken': page_token, 'Page': page_number, 'PageSize': page_size, })
+        data = values.of({
+            'Type': type,
+            'PageToken': page_token,
+            'Page': page_number,
+            'PageSize': page_size,
+        })
 
         response = self._version.page(method='GET', uri=self._uri, params=data, )